Correction of type for ID.

+++ src/create.c
@@ -63,8 +63,8 @@
 /* ---------------------------------------------------------------------------
  * some local identifiers
  */
-static int ID = 1;		/* current object ID; incremented after */
-				/* each creation of an object */
+static long int ID = 1;
+  /*!< Current object ID; incremented after each creation of an object */
 
 static bool be_lenient = false;
